(core): Reduce memory leak caused by continuously opening card dialogs
By adding class scope, we don't need to allocate memory every time the user opens the card dialog. The approach taken is to define a CardDetailsDialog instance as a member of ProgressWindow and have a CardWidget get the root (ProgressWindow) and open the dialog. The functionality is as equivalent as the later implementation.

The memory leak is still in investigation since the approach taken by the app is quite new. Also GTK is known to quite lazy when dealing with memory pages