(core): Change how Boards are interacted with
In later versions, Board instances were saved by utilising a BoardBackend helper class.
The overall problem was that the BoardBackend code interface at the time was quite clunky
and error-prone, thus doing the structure all over was necessary. The design approach chosen was
to provide a `Manager` that manages all kind of boards and keep track of them.

Another noticeable change in the core's structure is that we use signal to extend functionality. That
way, a code pattern for interacting with Progress's core is defined thus making easy to change
functionalities or add more features.

* Add BoardManager class
* Refactor Board, CardList, Card and Task classes to use sigc::signal<T(...)>
* Add sigc++ as a core dependency
* Rewrite Tests