(core): Delegate board loading to a thread
Loading boards linearly will get progressively slow as the
number of boards grow. Because there is no alternative known yet,
delegate board loading to a detached thread. Therefore, BoardManager constructor
will be a non-blocking operation in which clients must call to loaded
to ensure all boards in the database is loaded.

The methods are not (yet) thread-safe, so clients must synchronise the calls manually