(core): Initial app decoupling
The standard version of Progress delegates some of the application's
functionalities to its widgets, however, the problem with this approach is the
incorrect assumption that the widget's functionality won't ever be adapted, which
it's not the case. Thus this commit applies the first trial commit to decoupling the
core <-> widget relationship.