(chore): Simplify build process
We now use -DDEVELOPMENT flag to differ Debug builds from Release builds.
The "BUILD_TESTS" is also unnecessary since it is expected that setting
development mode means the developer expects the tests to be built as well.

Update build instructions as the last commit forgot to consider that the
project depends on submodules.