(widget): Create CoverColor helper enum
At the moment of this commit, I am still not satisfied with the current implementation.
This commit is meant as workaround for the fact we were using strings as keys, which
is not advisable. This is a signal that the current core implementation does not make
sense anymore, and it also deserves a refactor.